% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilutionplate_function_V3.R
\name{plate_map_extract}
\alias{plate_map_extract}
\title{Generate a 96-well plate map}
\usage{
plate_map_extract(ids, wtfile = TRUE, path = getwd(), des = "site",
  partial = T)
}
\arguments{
\item{ids}{Vector: sample identifiers (IDs)}

\item{wtfile}{Logical: Write 96-well plate templates to file in csv format}

\item{path}{Character: File directory where plate templates should be saved}

\item{des}{Character: Project description which will serve as a file prefix for all plates.}

\item{partial}{Logical: Designating if you want partial plates generated or not.}
}
\value{
List: 1st object is a list of plate matrices generated, 2nd object is list of plated samples, thier associated plate, and thier well coordinate. The 3rd object is any samples not plated when the argument partial is false.
}
\description{
This function takes a vector of sample identifiers and creates an arbitrary number of 96 well plates.
}
\examples{
samp_IDs<-paste("ind",1:1000,sep = "-")
plate_map_extract(ids = samp_IDs,wtfile = F,des = "Project_1",path = getwd(), partial = T)
}
\author{
Zak Robinson, Contact: zachary.robinson(at)umontana.com
}
